test_arrow_block
