+++ python/ray/data/tests/test_arrow_block.py
@@ -1,27 +1,18 @@
-import base64
 import os
 import sys
 import types
 from tempfile import TemporaryDirectory
-from typing import Union
 
 import numpy as np
 import pandas as pd
 import pyarrow as pa
 import pytest
-from pyarrow import ArrowInvalid
 
 import ray
 from ray._common.test_utils import run_string_as_driver
 from ray.data._internal.arrow_block import (
     ArrowBlockAccessor,
     ArrowBlockBuilder,
-    ArrowBlockColumnAccessor,
-    _get_max_chunk_size,
-)
-from ray.data._internal.arrow_ops.transform_pyarrow import combine_chunked_array, concat
-from ray.data._internal.tensor_extensions.arrow import (
-    ArrowTensorArray,
 )
 from ray.data._internal.util import GiB, MiB
 from ray.data.block import BlockAccessor
@@ -35,116 +26,6 @@ def simple_array():
 
 def simple_chunked_array():
     return pa.chunked_array([pa.array([1, 2]), pa.array([None, 6])])
-
-
-def _wrap_as_pa_scalar(v, dtype: pa.DataType):
-    return pa.scalar(v, type=dtype)
-
-
-@pytest.mark.parametrize("arr", [simple_array(), simple_chunked_array()])
-@pytest.mark.parametrize("as_py", [True, False])
-class TestArrowBlockColumnAccessor:
-    @pytest.mark.parametrize(
-        "ignore_nulls, expected",
-        [
-            (True, 3),
-            (False, 4),
-        ],
-    )
-    def test_count(self, arr, ignore_nulls, as_py, expected):
-        accessor = ArrowBlockColumnAccessor(arr)
-        result = accessor.count(ignore_nulls=ignore_nulls, as_py=as_py)
-
-        if not as_py:
-            expected = _wrap_as_pa_scalar(expected, dtype=pa.int64())
-
-        assert result == expected
-
-    @pytest.mark.parametrize(
-        "ignore_nulls, expected",
-        [
-            (True, 9),
-            (False, None),
-        ],
-    )
-    def test_sum(self, arr, ignore_nulls, as_py, expected):
-        accessor = ArrowBlockColumnAccessor(arr)
-        result = accessor.sum(ignore_nulls=ignore_nulls, as_py=as_py)
-
-        if not as_py:
-            expected = _wrap_as_pa_scalar(expected, dtype=pa.int64())
-
-        assert result == expected
-
-    @pytest.mark.parametrize(
-        "ignore_nulls, expected",
-        [
-            (True, 1),
-            (False, None),
-        ],
-    )
-    def test_min(self, arr, ignore_nulls, as_py, expected):
-        accessor = ArrowBlockColumnAccessor(arr)
-        result = accessor.min(ignore_nulls=ignore_nulls, as_py=as_py)
-
-        if not as_py:
-            expected = _wrap_as_pa_scalar(expected, dtype=pa.int64())
-
-        assert result == expected
-
-    @pytest.mark.parametrize(
-        "ignore_nulls, expected",
-        [
-            (True, 6),
-            (False, None),
-        ],
-    )
-    def test_max(self, arr, ignore_nulls, as_py, expected):
-        accessor = ArrowBlockColumnAccessor(arr)
-        result = accessor.max(ignore_nulls=ignore_nulls, as_py=as_py)
-
-        if not as_py:
-            expected = _wrap_as_pa_scalar(expected, dtype=pa.int64())
-
-        assert result == expected
-
-    @pytest.mark.parametrize(
-        "ignore_nulls, expected",
-        [
-            (True, 3),
-            (False, None),
-        ],
-    )
-    def test_mean(self, arr, ignore_nulls, as_py, expected):
-        accessor = ArrowBlockColumnAccessor(arr)
-        result = accessor.mean(ignore_nulls=ignore_nulls, as_py=as_py)
-
-        if not as_py:
-            expected = _wrap_as_pa_scalar(expected, dtype=pa.float64())
-
-        assert result == expected
-
-    @pytest.mark.parametrize(
-        "provided_mean, expected",
-        [
-            (3.0, 14.0),
-            (None, 14.0),
-        ],
-    )
-    def test_sum_of_squared_diffs_from_mean(self, arr, provided_mean, as_py, expected):
-        accessor = ArrowBlockColumnAccessor(arr)
-        result = accessor.sum_of_squared_diffs_from_mean(
-            ignore_nulls=True, mean=provided_mean, as_py=as_py
-        )
-
-        if not as_py:
-            expected = _wrap_as_pa_scalar(expected, dtype=pa.float64())
-
-        assert result == expected
-
-    def test_to_pylist(self, arr, as_py):
-        accessor = ArrowBlockColumnAccessor(arr)
-        assert accessor.to_pylist() == arr.to_pylist()
 
 
 @pytest.fixture(scope="module")
@@ -208,160 +89,6 @@ def test_single_row_gt_2gb(
     total = ds.map(_id).count()
 
     assert total == 1
-
-
-@pytest.mark.parametrize(
-    "input_,expected_output",
-    [
-        # Empty chunked array
-        (pa.chunked_array([], type=pa.int8()), pa.array([], type=pa.int8())),
-        # Fixed-shape tensors
-        (
-            pa.chunked_array(
-                [
-                    ArrowTensorArray.from_numpy(np.arange(3).reshape(3, 1)),
-                    ArrowTensorArray.from_numpy(np.arange(3).reshape(3, 1)),
-                ]
-            ),
-            ArrowTensorArray.from_numpy(
-                np.concatenate(
-                    [
-                        np.arange(3).reshape(3, 1),
-                        np.arange(3).reshape(3, 1),
-                    ]
-                )
-            ),
-        ),
-        # Ragged (variable-shaped) tensors
-        (
-            pa.chunked_array(
-                [
-                    ArrowTensorArray.from_numpy(np.arange(3).reshape(3, 1)),
-                    ArrowTensorArray.from_numpy(np.arange(5).reshape(5, 1)),
-                ]
-            ),
-            ArrowTensorArray.from_numpy(
-                np.concatenate(
-                    [
-                        np.arange(3).reshape(3, 1),
-                        np.arange(5).reshape(5, 1),
-                    ]
-                )
-            ),
-        ),
-        # Small (< 2 GiB) arrays
-        (
-            pa.chunked_array(
-                [
-                    pa.array([1, 2, 3], type=pa.int16()),
-                    pa.array([4, 5, 6], type=pa.int16()),
-                ]
-            ),
-            pa.array([1, 2, 3, 4, 5, 6], type=pa.int16()),
-        ),
-    ],
-)
-def test_combine_chunked_array_small(
-    input_, expected_output: Union[pa.Array, pa.ChunkedArray]
-):
-    result = combine_chunked_array(input_)
-
-    expected_output.equals(result)
-
-
-def test_combine_chunked_fixed_width_array_large():
-    """Verifies `combine_chunked_array` on fixed-width arrays > 2 GiB, produces
-    single contiguous PA Array"""
-
-    # 144 MiB
-    ones_1gb = np.ones(shape=(550, 128, 128, 4), dtype=np.int32()).ravel()
-
-    # Total ~2.15 GiB
-    input_ = pa.chunked_array(
-        [
-            pa.array(ones_1gb),
-        ]
-        * 16
-    )
-
-    assert round(input_.nbytes / GiB, 2) == 2.15
-
-    result = combine_chunked_array(input_)
-
-    assert isinstance(result, pa.Int32Array)
-
-
-@pytest.mark.parametrize(
-    "array_type,input_factory",
-    [
-        (
-            pa.binary(),
-            lambda num_bytes: np.arange(num_bytes, dtype=np.uint8).tobytes(),
-        ),
-        (
-            pa.string(),
-            lambda num_bytes: base64.encodebytes(
-                np.arange(num_bytes, dtype=np.int8).tobytes()
-            ).decode("ascii"),
-        ),
-        (pa.list_(pa.uint8()), lambda num_bytes: np.arange(num_bytes, dtype=np.uint8)),
-    ],
-)
-def test_combine_chunked_variable_width_array_large(array_type, input_factory):
-    """Verifies `combine_chunked_array` on variable-width arrays > 2 GiB,
-    safely produces new ChunkedArray with provided chunks recombined into
-    larger ones up to INT32_MAX in size"""
-
-    one_half_gb_arr = pa.array([input_factory(GiB / 2)], type=array_type)
-    chunked_arr = pa.chunked_array(
-        [one_half_gb_arr, one_half_gb_arr, one_half_gb_arr, one_half_gb_arr]
-    )
-
-    # 2 GiB + offsets (4 x int32)
-    num_bytes = chunked_arr.nbytes
-    expected_num_bytes = 4 * one_half_gb_arr.nbytes
-
-    num_chunks = len(chunked_arr.chunks)
-    assert num_chunks == 4
-    assert num_bytes == expected_num_bytes
-
-    # Assert attempt to combine directly fails
-    with pytest.raises(ArrowInvalid):
-        chunked_arr.combine_chunks()
-
-    # Safe combination succeeds by avoiding overflowing combination
-    combined = combine_chunked_array(chunked_arr)
-
-    num_bytes = combined.nbytes
-
-    num_chunks = len(combined.chunks)
-    assert num_chunks == 2
-    assert num_bytes == expected_num_bytes
-
-
-@pytest.mark.parametrize(
-    "input_block, fill_column_name, fill_value, expected_output_block",
-    [
-        (
-            pa.Table.from_pydict({"a": [0, 1]}),
-            "b",
-            2,
-            pa.Table.from_pydict({"a": [0, 1], "b": [2, 2]}),
-        ),
-        (
-            pa.Table.from_pydict({"a": [0, 1]}),
-            "b",
-            pa.scalar(2),
-            pa.Table.from_pydict({"a": [0, 1], "b": [2, 2]}),
-        ),
-    ],
-)
-def test_fill_column(input_block, fill_column_name, fill_value, expected_output_block):
-    block_accessor = ArrowBlockAccessor.for_block(input_block)
-
-    actual_output_block = block_accessor.fill_column(fill_column_name, fill_value)
-
-    assert actual_output_block.equals(expected_output_block)
 
 
 def test_random_shuffle(ray_start_regular_shared):
@@ -468,32 +195,6 @@ def test_build_block_with_null_column(ray_start_regular_shared):
     assert np.array_equal(rows[1]["array"], np.zeros((2, 2)))
 
 
-def test_add_rows_with_different_column_names():
-    builder = ArrowBlockBuilder()
-
-    builder.add({"col1": "spam"})
-    builder.add({"col2": "foo"})
-    block = builder.build()
-
-    expected_table = pa.Table.from_pydict(
-        {"col1": ["spam", None], "col2": [None, "foo"]}
-    )
-    assert block.equals(expected_table)
-
-
-def test_add_blocks_with_different_column_names():
-    builder = ArrowBlockBuilder()
-
-    builder.add_block(pa.Table.from_pydict({"col1": ["spam"]}))
-    builder.add_block(pa.Table.from_pydict({"col2": ["foo"]}))
-    block = builder.build()
-
-    expected_table = pa.Table.from_pydict(
-        {"col1": ["spam", None], "col2": [None, "foo"]}
-    )
-    assert block.equals(expected_table)
-
-
 def test_arrow_block_timestamp_ns(ray_start_regular_shared):
     # Input data with nanosecond precision timestamps
     data_rows = [
@@ -520,44 +221,20 @@ def test_arrow_block_timestamp_ns(ray_start_regular_shared):
         ), f"Timestamp mismatch at row {i} in ArrowBlockBuilder output"
 
 
-@pytest.mark.parametrize(
-    "table_data,max_chunk_size_bytes,expected",
-    [
-        ({"a": []}, 100, None),
-        ({"a": list(range(100))}, 7, 1),
-        ({"a": list(range(100))}, 10, 1),
-        ({"a": list(range(100))}, 25, 3),
-        ({"a": list(range(100))}, 50, 6),
-        ({"a": list(range(100))}, 100, 12),
-    ],
-)
-def test_arrow_block_max_chunk_size(table_data, max_chunk_size_bytes, expected):
-    table = pa.table(table_data)
-    assert _get_max_chunk_size(table, max_chunk_size_bytes) == expected
-
-
-def test_arrow_block_concat():
-    table1 = pa.table(
-        {
-            "a": [1, 2, 3],
-            "s": [{"x": 1} for _ in range(3)],
-        }
+def test_arrow_nan_element():
+    ds = ray.data.from_items(
+        [
+            1.0,
+            1.0,
+            2.0,
+            np.nan,
+            np.nan,
+        ]
     )
-    table2 = pa.table(
-        {
-            "b": [4, 5, 6],
-        }
-    )
-    concatenated = concat([table1, table2])
-    assert set(concatenated.column_names) == {"a", "s", "b"}
-    expected = pa.table(
-        {
-            "a": [1, 2, 3, None, None, None],
-            "s": [{"x": 1} for _ in range(3)] + [None] * 3,
-            "b": [None, None, None, 4, 5, 6],
-        }
-    )
-    assert concatenated.select(["a", "s", "b"]) == expected
+    ds = ds.groupby("item").count()
+    ds = ds.filter(lambda v: np.isnan(v["item"]))
+    result = ds.take_all()
+    assert result[0]["count()"] == 2
 
 
 if __name__ == "__main__":
